requirements.txt
argparse
PyYAML
keyboard
requests
pyjwt
termcolor
tabulate
tqdm
requests-toolbelt
beautifulsoup4
chardet
urllib3
